+++ platforms/cuda/src/CudaPyTorchKernelsE2EDiffConf.cpp
@@ -97,7 +97,7 @@ void CudaCalcPyTorchForceE2EDiffConfKernel::initialize(const System& system, con
 	vector<vector<int>> tmpPairs = force.getPairs();
 	vector<vector<int>> tmpTetras = force.getTetras();
 	vector<vector<int>> tmpCisTrans = force.getCisTrans();
-	vector<vector<float>> tmpEncoding = force.getEncoding();
+	vector<vector<double>> tmpEncoding = force.getEncoding();
 	usePeriodic = force.usesPeriodicBoundaryConditions();
 
 		
@@ -201,7 +201,7 @@ void CudaCalcPyTorchForceE2EDiffConfKernel::initialize(const System& system, con
 	// encoding
 	for (int i = 0; i < tmpEncoding.size(); i++) {
 		for (int j = 0; j < tmpEncoding[i].size(); j++) {
-			enc_acc[i][j] = tmpEncoding[i][j];
+		  enc_acc[i][j] = float(tmpEncoding[i][j]);
 		}
 	}
